backend/requirements.txt
groq>=0.11.0
sentence-transformers>=3.0.0
numpy>=1.24.0
pypdf>=4.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
httpx>=0.27.0
